requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
Django==5.0.4
django-environ==0.11.2
django-filter==24.2
django-parler==2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
inflection==0.5.1
moviepy==1.0.3
numpy==1.26.4
packaging==24.0
phonenumberslite==8.13.34
Pillow==9.5.0
proglog==0.1.10
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.2
uritemplate==4.1.1
urllib3==2.2.1
